+++ src/mrc.h
@@ -136,7 +136,7 @@ namespace Mrc {
         fseek(fp,92,SEEK_SET);
         fread((void*)(&offset),sizeof(uint32),1,fp);
         fseek(fp,1024+offset,SEEK_SET);
-        size_t num_el = X*Y*Z;
+        size_t num_el = (size_t)X*Y*Z;
         size_t read_el = fread((void*)buffer,sizeof(single),num_el,fp);
         fclose(fp);
         if( num_el != read_el ) {
@@ -147,7 +147,7 @@ namespace Mrc {
 
     float *read(uint32&X, uint32&Y, uint32&Z, const char*mapname) {
         read_size(X,Y,Z,mapname);
-        single *rslt = new single[X*Y*Z];
+        single *rslt = new single[(size_t)X*Y*Z];
         read(rslt,X,Y,Z,mapname);
         return rslt;
     }
@@ -178,7 +178,7 @@ namespace Mrc {
         header[53] = 0x00004444;
         if(fill_statistics) {
             float stats[4];
-            Math::get_min_max_avg_std(stats[0],stats[1],stats[2],stats[3],data,X*Y*Z);
+            Math::get_min_max_avg_std(stats[0],stats[1],stats[2],stats[3],data,(size_t)X*Y*Z);
             uint32*tmp = (uint32*)stats;
             header[19] = tmp[0];
             header[20] = tmp[1];
@@ -186,7 +186,7 @@ namespace Mrc {
             header[54] = tmp[3];
         }
         fwrite((void*)header,sizeof(uint32),256,fp);
-        fwrite((void*)data,sizeof(single),X*Y*Z,fp);
+        fwrite((void*)data,sizeof(single),(size_t)X*Y*Z,fp);
         fclose(fp);
     }
